fix test names; rename long constants; move buttons to appropriate file;
