+++ ui_tests_playwright/automation/identifiers/MainSite/MainSiteButtons.ts
@@ -20,4 +20,11 @@ export default class MainSiteButtons {
 	// Links
 	static Tiktok = 'Link-TikTok';
 	static Instagram = 'Link-Instagram';
+
+	static Technology_BackEnd = 'Button-BackEnd';
+	static Technology_FrontEnd = 'Button-FrontEnd';
+	static Technology_Mobile = 'Button-Mobile';
+	static Technology_Iot = 'Button-Iot';
+	static Technology_DevopsCloud = 'Button-DevopsCloud';
+	static Technology_AiMlDataScience = 'Button-AiMlDataScience';
 }
