+++ ui_tests_playwright/automation/identifiers/MainSite/pages/services/CustomDev.ts
@@ -1,14 +1,14 @@
 export class CustomDev {
-    static Info = 'Container-Info';
-    static CustomSoftwareDevelopmentForYourProduct = 'Container-CustomSoftwareDevelopmentForYourProduct';
-    static CustomSoftwareDevelopmentServicesWeProvide = 'Container-CustomSoftwareDevelopmentServicesWeProvide';
-    static TechnologyStack = 'Container-TechnologyStack';
-    static CustomSoftwareDevelopmentBenefits = 'Container-CustomSoftwareDevelopmentBenefits';
-    static CustomSoftwareDevelopmentProcess = 'Container-CustomSoftwareDevelopmentProcess';
-    static CustomSoftwareDevelopmentExperts = 'Container-CustomSoftwareDevelopmentExperts';
-    static CaseStudies = 'Container-CaseStudies';
-    static OurApproachToSoftwareDevelopment = 'Container-OurApproachToSoftwareDevelopment';
-    static Faq = 'Container-Faq';
-    static RelatedArticles = 'Container-RelatedArticles';
-    static GetInTouch = 'Container-GetInTouch';
+	static Info = 'Container-Info';
+	static CustomDevelopmentForYourProduct = 'Container-CustomSoftwareDevelopmentForYourProduct';
+	static CustomDevelopmentServicesWeProvide = 'Container-CustomSoftwareDevelopmentServicesWeProvide';
+	static TechnologyStack = 'Container-TechnologyStack';
+	static CustomDevelopmentBenefits = 'Container-CustomSoftwareDevelopmentBenefits';
+	static CustomDevelopmentProcess = 'Container-CustomSoftwareDevelopmentProcess';
+	static CustomDevelopmentExperts = 'Container-CustomSoftwareDevelopmentExperts';
+	static CaseStudies = 'Container-CaseStudies';
+	static OurApproachToSoftwareDevelopment = 'Container-OurApproachToSoftwareDevelopment';
+	static Faq = 'Container-Faq';
+	static RelatedArticles = 'Container-RelatedArticles';
+	static GetInTouch = 'Container-GetInTouch';
 }
